Factor out DeviceScreenService
